Use fixed 16 bit integers for compatibility between boards

+++ EasyTransferI2C/examples/EasyTransfer_RX_Example/EasyTransfer_RX_Example.pde
@@ -7,8 +7,8 @@ EasyTransferI2C ET;
 struct RECEIVE_DATA_STRUCTURE{
   //put your variable definitions here for the data you want to receive
   //THIS MUST BE EXACTLY THE SAME ON THE OTHER ARDUINO
-  int blinks;
-  int pause;
+  int16_t blinks;
+  int16_t pause;
 };
 
 //give a name to the group of data
